+++ tests/mocks/diagnostics-mock.hpp
@@ -14,7 +14,7 @@ class DiagnosticsMock : public ocls::IDiagnostics
 public:
     MOCK_METHOD(void, SetBuildOptions, (const nlohmann::json&), (override));
 
-    MOCK_METHOD(void, SetMaxProblemsCount, (int), (override));
+    MOCK_METHOD(void, SetMaxProblemsCount, (uint64_t), (override));
 
     MOCK_METHOD(void, SetOpenCLDevice, (uint32_t), (override));
